feat: implement PP store for the HTTP worker (#240)
